requirements.txt
langchain>=0.2.0
langgraph>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
elasticsearch==8.11.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic>=2.0.0
folium==0.15.0
jinja2==3.1.2
pyyaml==6.0.1
pandas>=2.0.0